dispositivos separados y sus metodos

+++ tp_integrador/tp_integrador/Models/Categoria.cs
@@ -8,11 +8,11 @@ namespace tp_integrador.Models
     public class Categoria
     {
         public string IdCategoria { get; set; }
-        public int Consumo { get; set; }
+        public byte Consumo { get; set; }
         public decimal CargoFijo { get; set; }
         public decimal CargoVariable { get; set; }
 
-        public Categoria(string id, int valor, decimal fijo, decimal variable)
+        public Categoria(string id, byte valor, decimal fijo, decimal variable)
         {
             IdCategoria = id;
             Consumo = valor;
